pyproject.toml
[tool.poetry]
name = "django-event-management"
version = "0.1.0"
description = "It is Django-based REST-Api that manages events (like conferences, meetups, etc.). The application will allow users to create, view, update, and delete events. It should also handle user registrations for these events."
authors = ["YuriiKrav4enko <yuriy.krav4enko28@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
django-allauth = "^64.2.0"
dj-rest-auth = "^6.0.0"
requests = "^2.32.3"
django-anymail = {extras = ["mailgun"], version = "^11.1"}
sentry-sdk = {extras = ["django"], version = "^2.13.0"}
django-templated-email = "^3.0.1"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
drf-spectacular = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
